Finished early set up of BFS method
Couple of things to note: 1. I forgot to add an exclamation mark at the start of the while loop, so it wasn't looping properly. 2. I implemented the logic for when we reach the target node (functionality not fully working yet, as "result" needs to be an array. 3. I changed some basic layout