Currently, established "numV" and "nodeIndex"
So numV represents the number of values/nodes similar to DSA lecture implementation. nodeIndex represents a mapping of each node to a value to be called upon. Then, we iterate over numV and assign each node a value (hence being able to call on nodes by their indices, not just the node itself)