Fixed my last issue with the code not passing tests
The second Edge test, getEdge2, was failing. I realised the issue here was that the if statement did not have the correct conditions. Initially, I was only checking that the edge's toNode was the destination, but no check that the edge came from the source desired. Therefore, it could have gotten the incorrect edge